[40] Remove hardcode secret
